Restart

+++ Assets/Scripts/EnergyBall.cs
@@ -6,14 +6,14 @@ using UnityEngine;
 public class EnergyBall : BaseBall
 {
     [SerializeField] private EnergyBar energy;
-    [SerializeField] protected int energyChange;
+    [SerializeField] protected float energyChange;
 
     protected override void Act()
     {
         energy.ChangeValue(GetEnergyChange());
     }
 
-    protected virtual int GetEnergyChange()
+    protected virtual float GetEnergyChange()
     {
         throw new NotImplementedException();
     }
